Allow option of `mongomock` for prod workloads

+++ src/argrelay/mongo_data/MongoServerWrapper.py
@@ -6,6 +6,7 @@ from argrelay.mongo_data.MongoConfig import MongoConfig
 
 
 class MongoServerWrapper:
+    is_mongomock: bool
     is_started: bool
     mongo_proc: Popen
 
@@ -13,8 +14,13 @@ class MongoServerWrapper:
         self.is_started = False
 
     def start_mongo_server(self, mongo_config: MongoConfig):
-        if mongo_config.start_server and not self.is_started:
-            self.mongo_proc = subprocess.Popen(mongo_config.server_start_command, shell = True)
+
+        if mongo_config.use_mongomock_only:
+            self.is_mongomock = True
+            return
+
+        if mongo_config.mongo_server.start_server and not self.is_started:
+            self.mongo_proc = subprocess.Popen(mongo_config.mongo_server.server_start_command, shell = True)
             try:
                 self.mongo_proc.wait(timeout = 5)
             except TimeoutExpired:
@@ -30,6 +36,10 @@ class MongoServerWrapper:
                 raise RuntimeError
 
     def stop_mongo_server(self):
+
+        if self.is_mongomock:
+            return
+
         if self.is_started:
             self.mongo_proc.kill()
         while self.is_started:
